Removing initial version of scripts, unifying with the new ones
